requirement.txt
selenium
webdriver_manager
requests
pyyaml
openpyxl